requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
openai>=1.12.0
tenacity>=8.2.0
python-dotenv>=1.0.0
google-genai>=0.1.0
numpy>=1.26.0